WIP: continued refactoring
